backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
appwrite==4.1.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
websockets>=12.0

# AI Agents
groq>=0.4.0
langchain>=0.1.0
langchain-groq>=0.0.3
google-generativeai>=0.4.0

# Media & TTS
Pillow>=10.0.0
elevenlabs>=1.0.0

# AWS S3
boto3>=1.34.0

# News harvesting
feedparser>=6.0.0
httpx>=0.26.0

# Scheduling & reliability
apscheduler>=3.10.0
tenacity>=8.2.0
